+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @amboni/tokens
 
+## 0.2.1
+
 ## 0.2.0
 
 ### Minor Changes